+++ test/spec/adapters/indexExchange_request_spec.js
@@ -4,7 +4,7 @@ import adLoader from '../../../src/adloader';
 
 var assert           = require('chai').assert;
 var IndexUtils       = require('../../helpers/index_adapter_utils.js');
-var HeaderTagRequest = '/headertag';
+var HeaderTagRequest = '/cygnus';
 var SlotThreshold    = 20;
 var ADAPTER_CODE     = 'indexExchange';
